azure-collector/azure-collector/requirements.txt
azure-identity
azure-mgmt-compute
azure-mgmt-network
neoj4